requirements.txt
# OllamaAid - Dependencies
# Core (required)
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# GUI (optional - install with: pip install ollama-aid[gui])
PySide6>=6.5.0

# Web (optional - install with: pip install ollama-aid[web])
flask>=3.0.0

# Tester extras (optional - install with: pip install ollama-aid[tester])
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0

# Screenshot generation (optional)
Pillow>=9.0.0

# Testing
pytest>=7.0
